sekretar - godisnji odmor lekar, pocetak validacije

+++ KT3/Projekat/Model/Pacijent.cs
@@ -24,7 +24,7 @@ namespace Model
     }
     public class Pacijent : INotifyPropertyChanged
     {
-        public Pacijent(int IdPacijenta, string ImePacijenta, string PrezimePacijenta, int Jmbg, pol pol, long BrojTelefona, string Email, string AdresaStanovanja, statusNaloga Status, String zanimanje, bracnoStanje stanje)
+        public Pacijent(int IdPacijenta, string ImePacijenta, string PrezimePacijenta, long Jmbg, pol pol, long BrojTelefona, string Email, string AdresaStanovanja, statusNaloga Status, String zanimanje, bracnoStanje stanje)
         {
             this.IdPacijenta = IdPacijenta;
             this.ImePacijenta = ImePacijenta;
@@ -39,7 +39,7 @@ namespace Model
             this.BracnoStanje = stanje;
         }
 
-        public Pacijent(int IdPacijenta, string ImePacijenta, string PrezimePacijenta, int Jmbg, pol pol, long BrojTelefona, string Email, string AdresaStanovanja, statusNaloga Status, String zanimanje, bracnoStanje stanje, bool Maloletnik, int JmbgStaratelja)
+        public Pacijent(int IdPacijenta, string ImePacijenta, string PrezimePacijenta, long Jmbg, pol pol, long BrojTelefona, string Email, string AdresaStanovanja, statusNaloga Status, String zanimanje, bracnoStanje stanje, bool Maloletnik, long JmbgStaratelja)
         {
             this.IdPacijenta = IdPacijenta;
             this.ImePacijenta = ImePacijenta;
@@ -58,7 +58,7 @@ namespace Model
 
         public Pacijent() { }
 
-        public Pacijent(int IdPacijenta, string ImePacijenta, string PrezimePacijenta, int Jmbg, pol pol, statusNaloga Status)
+        public Pacijent(int IdPacijenta, string ImePacijenta, string PrezimePacijenta, long Jmbg, pol pol, statusNaloga Status)
         {
             this.IdPacijenta = IdPacijenta;
             this.ImePacijenta = ImePacijenta;
@@ -72,7 +72,7 @@ namespace Model
         public string ImePacijenta { get; set; }
         public string PrezimePacijenta { get; set; }
         public pol Pol { get; set; }
-        public int Jmbg { get; set; }
+        public long Jmbg { get; set; }
         public statusNaloga StatusNaloga { get; set; }
         public long BrojTelefona { get; set; }
         public string Email { get; set; }
@@ -82,7 +82,7 @@ namespace Model
         public Lekar IzabraniLekar { get; set; }
         public ZdravstveniKarton Karton { get; set; }
         public bool Maloletnik { get; set; }
-        public int JmbgStaratelja { get; set; } 
+        public long JmbgStaratelja { get; set; } 
 
         public event PropertyChangedEventHandler PropertyChanged;
         protected virtual void OnPropertyChanged(string name)
